requirements.txt
asgiref==3.7.2
Django==4.2.9
django-rest-framework==0.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
psycopg2-binary==2.9.7
